pyflink-payten/DE-payten/data-generators/requirements.txt
Faker==8.8.1
pandas==1.2.4
kafka-python==2.0.2
